+++ packages/react-native/ReactAndroid/build.gradle
@@ -383,6 +383,13 @@ task downloadNdkBuildDependencies {
     dependsOn(downloadGtest)
 }
 
+// As ReactAndroid builds from source, the codegen needs to be built before it can be invoked.
+// This is not the case for users of React Native, as we ship a compiled version of the codegen.
+final def buildCodegenCLITask = tasks.register('buildCodegenCLI', BuildCodegenCLITask) {
+    it.codegenDir.set(file("$rootDir/node_modules/@react-native/codegen"))
+    it.bashWindowsHome.set(project.findProperty("REACT_WINDOWS_BASH"))
+}
+
 /**
  * Finds the path of the installed npm package with the given name using Node's
  * module resolution algorithm, which searches "node_modules" directories up to
@@ -572,9 +579,20 @@ android {
         }
     }
 
-    preBuild.dependsOn(prepareJSC, prepareBoost, prepareDoubleConversion, prepareFmt, prepareFolly, prepareGlog, prepareLibevent, prepareGtest)
-    preBuild.dependsOn("generateCodegenArtifactsFromSchema")
-    preBuild.dependsOn(preparePrefab)
+    preBuild.dependsOn(
+            buildCodegenCLITask,
+            generateCodegenArtifactsFromSchema,
+            prepareBoost,
+            prepareDoubleConversion,
+            prepareFmt,
+            prepareFolly,
+            prepareGlog,
+            prepareGtest,
+            prepareJSC,
+            prepareLibevent,
+            preparePrefab
+    )
+    generateCodegenSchemaFromJavaScript.dependsOn(buildCodegenCLITask)
 
     sourceSets.main {
         res.srcDirs = ["src/main/res/devsupport", "src/main/res/shell", "src/main/res/views/modal", "src/main/res/views/uimanager"]
